+++ src/app/components/cluster-host-table-row.js
@@ -1,25 +1,95 @@
 import Component from '@ember/component';
 import { alias } from '@ember/object/computed';
-import {
-  InvokeActionMixin,
-} from 'ember-invoke-action';
+import notImplementedReject from 'onedata-gui-common/utils/not-implemented-reject';
+import notImplementedWarn from 'onedata-gui-common/utils/not-implemented-warn';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
+import { computed, get } from '@ember/object';
+import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
+import { inject as service } from '@ember/service';
 
-export default Component.extend(InvokeActionMixin, {
+export default Component.extend({
   tagName: 'tr',
-  classNames: 'cluster-host-table-row',
-  classNameBindings: ['active'],
+  classNames: ['cluster-host-table-row', 'animated', 'infinite'],
+  classNameBindings: ['active', 'blinking:pulse-bg-mint'],
   attributeBindings: ['dataHostname:data-hostname'],
+
+  onepanelServer: service(),
+
+  /**
+   * @virtual
+   * @type {ClusterHostInfo}
+   */
+  host: undefined,
+
+  /**
+   * @virtual
+   * @type {Function}
+   */
+  checkboxChanged: notImplementedWarn,
+
+  /**
+   * @virtual
+   * @type {Function} (hostname: string) => Promise
+   */
+  removeHost: notImplementedReject,
+
+  /**
+   * @virtual 
+   * @type {boolean}
+   */
+  isMobile: undefined,
+
+  removeAvailable: computed('removeHost', function () {
+    return this.get('removeHost') !== notImplementedReject;
+  }),
 
   /**
    * @type {boolean}
    */
   active: false,
 
-  // TODO security - check if cannot make HTML injection using hostname
+  /**
+   * @type {boolean}
+   */
+  blinking: false,
+
   /**
    * @type {ClusterHostInfo}
    */
   dataHostname: alias('host.hostname'),
+
+  /**
+   * @type {boolean}
+   */
+  _removeDisabled: false,
+
+  _removeHostBtnClass: computed('isThisHost.{isSettled,content}', function () {
+    const isThisHost = this.get('isThisHost');
+    if (get(isThisHost, 'isSettled')) {
+      if (get(isThisHost, 'isRejected')) {
+        return 'disabled';
+      } else {
+        return get(isThisHost, 'content') ? 'disabled' : '';
+      }
+    } else {
+      return 'disabled';
+    }
+  }),
+
+  /**
+   * True if this row represents current host that 
+   * @type {Ember.ComputedProperty<Promise<boolean>>}
+   */
+  isThisHost: computed('onepanelServer', 'dataHostname', function () {
+    const {
+      onepanelServer,
+      dataHostname,
+    } = this.getProperties('onepanelServer', 'dataHostname');
+    return PromiseObject.create({
+      promise: onepanelServer.getHostname()
+        .then(hostname => hostname === dataHostname),
+    });
+  }),
 
   actions: {
     headerClick() {
@@ -31,10 +101,19 @@ export default Component.extend(InvokeActionMixin, {
     ) {
       let hostname = context.get('hostHostname');
       let option = context.get('hostOption');
-      this.invokeAction('checkboxChanged', hostname, option, newValue);
+      this.get('checkboxChanged')(hostname, option, newValue);
     },
     primaryClusterManagerChanged() {
-      this.invokeAction('primaryClusterManagerChanged', ...arguments);
+      this.get('primaryClusterManagerChanged')(...arguments);
+    },
+    removeHost() {
+      const {
+        removeHost,
+        dataHostname,
+      } = this.getProperties('removeHost', 'dataHostname');
+      this.set('_removeDisabled', true);
+      return removeHost(dataHostname)
+        .finally(() => safeExec(this, 'set', '_removeDisabled', false));
     },
   },
 });