+++ src/app/locales/en/components/new-cluster-installation.js
@@ -3,4 +3,24 @@ export default {
   deploy: 'Deploy',
   zoneOptions: 'Zone options',
   clusterHosts: 'Cluster hosts',
+  domainLink: 'the documentation',
+  addNewHost: 'Add new host',
+  newHostnameExample: 'Enter hostname or IP address',
+  save: 'Add host',
+  cancel: 'Cancel',
+  addingNewHost: 'adding new host',
+  showAddInformation: 'Show instructions...',
+  hideAddInformation: 'Hide instructions',
+  addInfoStepsDescription: 'To add a new node to the cluster, follow these steps:',
+  addInfoStep1: 'Install {{serviceName}} panel on the node.',
+  addInfoStep2: 'Make sure that the following ports are open between cluster ' +
+    'nodes hosting certain services:',
+  addHostInfoClusterPorts: 'Cluster Worker or Cluster Manager service: ' +
+    '{{clusterPorts}}',
+  addHostInfoCouchbasePorts: 'Couchbase: {{couchbasePorts}}',
+  addInfoStep3: 'Insert the IP address or hostname of the node to add in the ' +
+    'above field, or visit the onepanel web interface of the node and press ' +
+    '"Join cluster" to easily obtain node hostname.',
+  addInfoStep4: 'If the new node is reachable and the ports are open, it will ' +
+    'be discovered and added to the cluster.',
 };